Set project up in Context
Moves sidebar and main frame into separate components. Instantiates
context and basic hook.
