--- src/layout/primary.js
@@ -1,77 +1,0 @@
-
-import { React, Component } from 'react'
-
-import logo from '../assets/mill_logo.png'
-import textLogo from '../assets/text-logo.png'
-
-import { Layout, Menu, Breadcrumb, Typography } from 'antd';
-
-import {
-  EnvironmentOutlined,
-  PieChartOutlined,
-  ScheduleOutlined,
-  ToolOutlined,
-} from '@ant-design/icons';
-
-import './primary.css'
-
-const { Title } = Typography;
-const { Header, Content, Footer, Sider } = Layout;
-
-class SiderDemo extends Component {
-  state = {
-    collapsed: false,
-    user: 'Ashley',
-    currentPage: 'Dashboard',
-  };
-
-  onCollapse = collapsed => {
-    this.setState({ collapsed });
-  };
-
-  render() {
-    const { collapsed } = this.state;
-    return (
-      <Layout style={{ minHeight: '100vh' }}>
-        <Sider className={`sidebar-frame ${this.state.collapsed && 'collapsed'}`} collapsible collapsed={collapsed} onCollapse={this.onCollapse} >
-          <div onClick={() => this.setState({currentPage: 'Dashboard' })} className={`logo ${this.state.collapsed && 'collapsed-logo'}`} >
-            <img src={logo} alt='Windmill logo'/>
-            {!this.state.collapsed && (<img className='text-logo' src={textLogo} alt='Skyline text logo'/>)}
-          </div>
-          <Menu theme="light" defaultSelectedKeys={['1']} mode="inline">
-            <Menu.Item key="1" icon={<EnvironmentOutlined />} onClick={() => this.setState({currentPage: 'Fleet'})} style={{ marginTop: '10px'}}>
-              Fleet
-            </Menu.Item>
-            <Menu.Item key="2" onClick={() => this.setState({currentPage: 'Inspections'})} icon={<ScheduleOutlined />}>
-              Inspections
-            </Menu.Item>
-            <Menu.Item key="10" onClick={() => this.setState({currentPage: 'Work Orders'})} icon={<ToolOutlined />}>
-              Work Orders
-            </Menu.Item>
-            <Menu.Item key="9" onClick={() => this.setState({currentPage: 'Analytics'})} icon={<PieChartOutlined />}>
-              Analytics
-            </Menu.Item>
-          </Menu>
-        </Sider>
-        <Layout className="site-layout">
-          <div><Title className="nav-selection-frame-title" level={2}>{this.state.currentPage}</Title>
-          <Breadcrumb className="user-info-display">
-              <Breadcrumb.Item>User</Breadcrumb.Item>
-              <Breadcrumb.Item>{this.state.user}</Breadcrumb.Item>
-            </Breadcrumb></div>
-          <Header className="frame-navigation" />
-          <Content style={{ margin: '0 16px' }}>
-            <div className="site-layout-background" style={{ padding: 24, minHeight: 360 }}>
-              This is a mock application being built in the style of an existing application, using promotional material as a model.
-            </div>
-          </Content>
-          <Footer style={{ textAlign: 'center' }}>UI built with Ant Design</Footer>
-        </Layout>
-      </Layout>
-    );
-  }
-}
-
-export default SiderDemo
-
-// END of document
